python_src/requirements.txt
libvirt-python==7.0.0
